Moved discover logic code from __init__ file to discover.py file
